+++ docs/content/docs/api/modals-component.mdx
@@ -2,8 +2,6 @@
 title: <Modals />
 layout: $lib/layouts/BaseLayout.astro
 ---
-import { Code } from '@astrojs/starlight/components';
-import ModalsDTS from '../../../../dist/modals-context.svelte.d.ts?raw'
 
 Renders the current stack of modals
