+++ docs/content/docs/api/modals.mdx
@@ -3,7 +3,7 @@ title: modals
 layout: $lib/layouts/BaseLayout.astro
 ---
 import { Code, Badge} from '@astrojs/starlight/components';
-import ModalsDTS from '../../../../dist/modals-context.svelte.d.ts?raw'
+import ModalsDTS from '../../../../dist/modal-stack.svelte.d.ts?raw'
 
 The main entry point for the modals API.
